Version is increased to support market data in shares (#1037)

+++ source/csharpclient/TwsRtdServer/TwsRtdServerMktDataTicks.cs
@@ -13,12 +13,12 @@ namespace TwsRtdServer
         private double m_lastPrice = 0.0;
         private double m_bidPrice = 0.0;
         private double m_askPrice = 0.0;
-        private int m_lastSize = 0;
-        private int m_bidSize = 0;
-        private int m_askSize = 0;
+        private long m_lastSize = 0;
+        private long m_bidSize = 0;
+        private long m_askSize = 0;
         private double m_high = 0.0;
         private double m_low = 0.0;
-        private int m_volume = 0;
+        private long m_volume = 0;
         private double m_close = 0.0;
         private double m_open = 0.0;
         private string m_lastExch = "";
@@ -34,17 +34,17 @@ namespace TwsRtdServer
         private TwsRtdServerData.OptionComputationData m_modelOptionComputation = new TwsRtdServerData.OptionComputationData();
 
         // add generic ticks values
-        private int m_genTickAuctionVolume = 0;
-        private int m_genTickAuctionImbalance = 0;
+        private long m_genTickAuctionVolume = 0;
+        private long m_genTickAuctionImbalance = 0;
         private double m_genTickAuctionPrice = 0.0;
-        private int m_genTickRegulatoryImbalance = 0;
+        private long m_genTickRegulatoryImbalance = 0;
         private double m_genTickPlPrice = 0.0;
         private double m_genTickCreditmanMarkPrice = 0.0;
         private double m_genTickCreditmanSlowMarkPrice = 0.0;
-        private int m_genTickCallOptionVolume = 0;
-        private int m_genTickPutOptionVolume = 0;
-        private int m_genTickCallOptionOpenInterest = 0;
-        private int m_genTickPutOptionOpenInterest = 0;
+        private long m_genTickCallOptionVolume = 0;
+        private long m_genTickPutOptionVolume = 0;
+        private long m_genTickCallOptionOpenInterest = 0;
+        private long m_genTickPutOptionOpenInterest = 0;
         private double m_genTickOptionHistoricalVol = 0.0;
         private double m_genTickRTHistoricalVol = 0.0;
         private double m_genTickIndexFuturePremium = 0.0;
@@ -56,19 +56,19 @@ namespace TwsRtdServer
         private double m_genTickLastRTHTrade = 0.0;
         private string m_genTickIBDividends = "";
         private double m_genTickBondMultiplier = 0.0;
-        private int m_genTickAverageVolume = 0;
+        private long m_genTickAverageVolume = 0;
         private double m_genTickWeek13Hi = 0.0;
         private double m_genTickWeek13Lo = 0.0;
         private double m_genTickWeek26Hi = 0.0;
         private double m_genTickWeek26Lo = 0.0;
         private double m_genTickWeek52Hi = 0.0;
         private double m_genTickWeek52Lo = 0.0;
-        private int m_genTickShortTermVolume3Min = 0;
-        private int m_genTickShortTermVolume5Min = 0;
-        private int m_genTickShortTermVolume10Min = 0;
-        private int m_futuresOpenInterest = 0;
-        private int m_genTickAvgOptVolume = 0;
-        private int m_genTickShortableShares = 0;
+        private long m_genTickShortTermVolume3Min = 0;
+        private long m_genTickShortTermVolume5Min = 0;
+        private long m_genTickShortTermVolume10Min = 0;
+        private long m_futuresOpenInterest = 0;
+        private long m_genTickAvgOptVolume = 0;
+        private long m_genTickShortableShares = 0;
         private double m_genTickEtfNavClose = 0;
         private double m_genTickEtfNavPriorClose = 0;
         private double m_genTickEtfNavBid = 0;
@@ -82,12 +82,12 @@ namespace TwsRtdServer
         private double m_delayed_lastPrice = 0.0;
         private double m_delayed_bidPrice = 0.0;
         private double m_delayed_askPrice = 0.0;
-        private int m_delayed_lastSize = 0;
-        private int m_delayed_bidSize = 0;
-        private int m_delayed_askSize = 0;
+        private long m_delayed_lastSize = 0;
+        private long m_delayed_bidSize = 0;
+        private long m_delayed_askSize = 0;
         private double m_delayed_high = 0.0;
         private double m_delayed_low = 0.0;
-        private int m_delayed_volume = 0;
+        private long m_delayed_volume = 0;
         private double m_delayed_close = 0.0;
         private double m_delayed_open = 0.0;
         private string m_delayed_lastTimestamp = "";
@@ -112,13 +112,13 @@ namespace TwsRtdServer
                     m_bidPrice = (double)value;
                     break;
                 case TwsRtdServerData.LASTSIZE:
-                    m_lastSize = (int)value;
+                    m_lastSize = (long)value;
                     break;
                 case TwsRtdServerData.ASKSIZE:
-                    m_askSize = (int)value;
+                    m_askSize = (long)value;
                     break;
                 case TwsRtdServerData.BIDSIZE:
-                    m_bidSize = (int)value;
+                    m_bidSize = (long)value;
                     break;
                 case TwsRtdServerData.HIGH:
                     m_high = (double)value;
@@ -136,7 +136,7 @@ namespace TwsRtdServer
                     m_halted = (double)value;
                     break;
                 case TwsRtdServerData.VOLUME:
-                    m_volume = (int)value;
+                    m_volume = (long)value;
                     break;
                 case TwsRtdServerData.BIDEXCH:
                     m_bidExch = (string)value;
@@ -164,16 +164,16 @@ namespace TwsRtdServer
                     break;
                 // generic ticks
                 case TwsRtdServerData.GEN_TICK_AUCTION_VOLUME:
-                    m_genTickAuctionVolume = (int)value;
+                    m_genTickAuctionVolume = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_AUCTION_IMBALANCE:
-                    m_genTickAuctionImbalance = (int)value;
+                    m_genTickAuctionImbalance = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_AUCTION_PRICE:
                     m_genTickAuctionPrice = (double)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_REGULATORY_IMBALANCE:
-                    m_genTickRegulatoryImbalance = (int)value;
+                    m_genTickRegulatoryImbalance = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_PL_PRICE:
                     m_genTickPlPrice = (double)value;
@@ -185,16 +185,16 @@ namespace TwsRtdServer
                     m_genTickCreditmanSlowMarkPrice = (double)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_CALL_OPTION_VOLUME:
-                    m_genTickCallOptionVolume = (int)value;
+                    m_genTickCallOptionVolume = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_PUT_OPTION_VOLUME:
-                    m_genTickPutOptionVolume = (int)value;
+                    m_genTickPutOptionVolume = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_CALL_OPTION_OPEN_INTEREST:
-                    m_genTickCallOptionOpenInterest = (int)value;
+                    m_genTickCallOptionOpenInterest = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_PUT_OPTION_OPEN_INTEREST:
-                    m_genTickPutOptionOpenInterest = (int)value;
+                    m_genTickPutOptionOpenInterest = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_OPTION_HISTORICAL_VOL:
                     m_genTickOptionHistoricalVol = (double)value;
@@ -230,7 +230,7 @@ namespace TwsRtdServer
                     m_genTickBondMultiplier = (double)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_AVERAGE_VOLUME:
-                    m_genTickAverageVolume = (int)value;
+                    m_genTickAverageVolume = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_WEEK_13_HI:
                     m_genTickWeek13Hi = (double)value;
@@ -251,22 +251,22 @@ namespace TwsRtdServer
                     m_genTickWeek52Lo = (double)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_SHORT_TERM_VOLUME_3_MIN:
-                    m_genTickShortTermVolume3Min = (int)value;
+                    m_genTickShortTermVolume3Min = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_SHORT_TERM_VOLUME_5_MIN:
-                    m_genTickShortTermVolume5Min = (int)value;
+                    m_genTickShortTermVolume5Min = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_SHORT_TERM_VOLUME_10_MIN:
-                    m_genTickShortTermVolume10Min = (int)value;
+                    m_genTickShortTermVolume10Min = (long)value;
                     break;
                 case TwsRtdServerData.FUTURES_OPEN_INTEREST:
-                    m_futuresOpenInterest = (int)value;
+                    m_futuresOpenInterest = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_AVG_OPT_VOLUME:
-                    m_genTickAvgOptVolume = (int)value;
+                    m_genTickAvgOptVolume = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_SHORTABLE_SHARES:
-                    m_genTickShortableShares = (int)value;
+                    m_genTickShortableShares = (long)value;
                     break;
                 case TwsRtdServerData.GEN_TICK_ETF_NAV_CLOSE:
                     m_genTickEtfNavClose = (double)value;
@@ -304,13 +304,13 @@ namespace TwsRtdServer
                     m_delayed_askPrice = (double)value;
                     break;
                 case TwsRtdServerData.DELAYED_LAST_SIZE:
-                    m_delayed_lastSize = (int)value;
+                    m_delayed_lastSize = (long)value;
                     break;
                 case TwsRtdServerData.DELAYED_BID_SIZE:
-                    m_delayed_bidSize = (int)value;
+                    m_delayed_bidSize = (long)value;
                     break;
                 case TwsRtdServerData.DELAYED_ASK_SIZE:
-                    m_delayed_askSize = (int)value;
+                    m_delayed_askSize = (long)value;
                     break;
                 case TwsRtdServerData.DELAYED_HIGH:
                     m_delayed_high = (double)value;
@@ -319,7 +319,7 @@ namespace TwsRtdServer
                     m_delayed_low = (double)value;
                     break;
                 case TwsRtdServerData.DELAYED_VOLUME:
-                    m_delayed_volume = (int)value;
+                    m_delayed_volume = (long)value;
                     break;
                 case TwsRtdServerData.DELAYED_CLOSE:
                     m_delayed_close = (double)value;
